Remove unnecessary ReportScreenWrapper component
